Rename files to use UpperCamelCaseCluster/Logic for webrtc-transport-requestor-server
