+++ src/app/clusters/webrtc-transport-requestor-server/BUILD.gn
@@ -16,8 +16,8 @@ import("//build_overrides/chip.gni")
 
 source_set("webrtc-transport-requestor-server") {
   sources = [
-    "webrtc-transport-requestor-cluster.cpp",
-    "webrtc-transport-requestor-cluster.h",
+    "WebRTCTransportRequestorCluster.cpp",
+    "WebRTCTransportRequestorCluster.h",
   ]
 
   public_deps = [
