+++ src/app/clusters/webrtc-transport-requestor-server/app_config_dependent_sources.cmake
@@ -16,6 +16,6 @@
 TARGET_SOURCES(
   ${APP_TARGET}
   PRIVATE
-    "${CLUSTER_DIR}/webrtc-transport-requestor-cluster.h"
-    "${CLUSTER_DIR}/webrtc-transport-requestor-cluster.cpp"
+    "${CLUSTER_DIR}/WebRTCTransportRequestorCluster.h"
+    "${CLUSTER_DIR}/WebRTCTransportRequestorCluster.cpp"
 )
